spawnprotected back in msec instead of sec
To avoid overridance from other mods that cause g_spawnprotect to be set to 500 seconds

+++ omega_gamecode/code/game/g_active.c
@@ -43,7 +43,7 @@ static void P_DamageFeedback( gentity_t *player ) {
 		return;
 	}
 
-	if ( level.time < client->respawnTime + g_spawnprotect.integer * 1000 && client->spawnprotected ) {
+	if ( level.time < client->respawnTime + g_spawnprotect.integer && client->spawnprotected ) {
 		client->ps.powerups[PW_BATTLESUIT] = -1;
 	}
